requirements-files.txt
# requirements/base.txt
# Core requirements for all nodes
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1
psutil==5.9.6
dataclasses-json==0.6.3
schedule==1.2.0
click==8.1.7
colorlog==6.8.0
tenacity==8.2.3

# requirements/ai.txt
# AI Node specific requirements
# Install base requirements first: pip install -r requirements/base.txt

# PyTorch (choose based on your CUDA version)
# For CUDA 11.8
torch==2.1.2+cu118 --index-url https://download.pytorch.org/whl/cu118
torchvision==0.16.2+cu118 --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1
# torch==2.1.2+cu121 --index-url https://download.pytorch.org/whl/cu121
# torchvision==0.16.2+cu121 --index-url https://download.pytorch.org/whl/cu121

# ML/AI libraries
transformers==4.36.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
scipy==1.11.4
accelerate==0.25.0
sentencepiece==0.1.99
safetensors==0.4.1

# NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1
vaderSentiment==3.3.2

# Video/Image processing
opencv-python==4.8.1.78
pillow==10.1.0
moviepy==1.0.3

# requirements/scraping.txt
# Scraping Node specific requirements
# Install base requirements first: pip install -r requirements/base.txt

# HTTP/Scraping
httpx==0.25.2
beautifulsoup4==4.12.2
lxml==4.9.4
html5lib==1.1

# Browser automation
selenium==4.16.0
undetected-chromedriver==3.5.4
playwright==1.40.0
selenium-stealth==1.0.6

# Anti-detection
fake-useragent==1.4.0
cloudscraper==1.2.71

# Platform specific
praw==7.7.1  # Reddit
snscrape==0.7.0.20230622  # Twitter
facebook-scraper==0.2.59  # Facebook
TikTokApi==6.2.0  # TikTok

# Proxy support
pysocks==1.7.1
proxy-requests==0.3.5

# requirements/master.txt
# View Master specific requirements
# Install base requirements first: pip install -r requirements/base.txt

# Web framework
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Monitoring
prometheus-client==0.19.0
grafana-api==1.0.3

# Visualization
plotly==5.18.0
dash==2.14.2

# .env.example
# ScrapeHive Configuration

# Redis Configuration
REDIS_HOST=localhost
REDIS_PORT=6379
REDIS_PASSWORD=
REDIS_DB=0

# Master Node
MASTER_IP=192.168.1.100
MASTER_PORT=5000

# Proxy Configuration (Optional)
PROXY_PROVIDER=brightdata
PROXY_USERNAME=your_username
PROXY_PASSWORD=your_password
PROXY_ENDPOINT=http://proxy.provider.com:8080

# AI Model Paths (for AI node)
MODEL_PATH=/path/to/models
WHISPER_MODEL=base
DEVICE=cuda  # or cpu

# Scraping Limits
REDDIT_RATE_LIMIT=30
TWITTER_RATE_LIMIT=20
INSTAGRAM_RATE_LIMIT=10
TIKTOK_RATE_LIMIT=15

# Database
DATABASE_PATH=~/.scrapehive/master.db

# Logging
LOG_LEVEL=INFO
LOG_PATH=~/.scrapehive/logs

# Security
SECRET_KEY=your-secret-key-here
API_KEY=your-api-key-for-external-access